--- src/main/java/cart/controller/dto/ExceptionResponse.java
@@ -1,14 +1,0 @@
-package cart.controller.dto;
-
-public class ExceptionResponse<T> {
-
-    private final T message;
-
-    public ExceptionResponse(T message) {
-        this.message = message;
-    }
-
-    public T getMessage() {
-        return message;
-    }
-}
