+++ src/main/java/cart/exception/dto/ExceptionResponse.java
@@ -1,0 +1,14 @@
+package cart.exception.dto;
+
+public class ExceptionResponse {
+
+    private final String message;
+
+    public ExceptionResponse(String message) {
+        this.message = message;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+}
